+++ api/src/main/java/me/combimagnetron/sunscreen/event/ClickElementEvent.java
@@ -4,10 +4,11 @@ import me.combimagnetron.passport.event.Event;
 import me.combimagnetron.sunscreen.element.Element;
 import me.combimagnetron.sunscreen.menu.input.Input;
 import me.combimagnetron.sunscreen.util.Vec2d;
+import me.combimagnetron.sunscreen.util.Vec2i;
 
-public record ClickElementEvent<T extends Element>(T element, Vec2d coords, Input.Type.MouseClick click) implements Event {
+public record ClickElementEvent<T extends Element>(T element, Vec2i coords, Input.Type.MouseClick click) implements Event {
 
-    public static <T extends Element> ClickElementEvent<T> create(T element, Vec2d coords, Input.Type.MouseClick click) {
+    public static <T extends Element> ClickElementEvent<T> create(T element, Vec2i coords, Input.Type.MouseClick click) {
         return new ClickElementEvent<>(element, coords, click);
     }
