+++ api/src/main/java/me/combimagnetron/sunscreen/renderer/div/ItemDivRenderer.java
@@ -20,7 +20,7 @@ public final class ItemDivRenderer implements DivRenderer<ItemDisplay> {
             itemDisplay.brightness(15, 15);
             itemDisplay.displayType(ItemDisplay.DisplayType.GUI);
             Display.Transformation transformation = Display.Transformation.transformation();
-            transformation = transformation.translation(ViewportHelper.toTranslation(ViewportHelper.fromPosition(div.position()).add(div.size().x() * 0.5 * div.scale().x(), div.size().y() * div.scale().y()), user.screenSize()).add(Vector3d.vec3(0, 0, -0.25 + Integer.MIN_VALUE * div.order())));
+            transformation = transformation.translation(ViewportHelper.toTranslation(ViewportHelper.fromPosition(div.position()).add((int) (div.size().x() * 0.5 * div.scale().x()), (int) (div.size().y() * div.scale().y())), user.screenSize()).add(Vector3d.vec3(0, 0, -0.25 + Integer.MIN_VALUE * div.order())));
             transformation = transformation.scale(Vector3d.vec3((double) 1 / 24).mul(div.scale()));
             itemDisplay.transformation(transformation);
             return referenceHolder.reference(itemDisplay, div);
