+++ api/src/main/java/me/combimagnetron/sunscreen/util/HoverHelper.java
@@ -6,16 +6,16 @@ import me.combimagnetron.sunscreen.user.SunscreenUser;
 
 public class HoverHelper {
 
-    public static boolean isHovered(Vector3d cursorTranslation, SunscreenUser<?> viewer, Vec2d position, Vec2d size) {
-        Vec2d cursor = ViewportHelper.toScreen(cursorTranslation, viewer.screenSize());
+    public static boolean isHovered(Vector3d cursorTranslation, SunscreenUser<?> viewer, Vec2i position, Vec2i size) {
+        Vec2i cursor = ViewportHelper.toScreen(cursorTranslation, viewer.screenSize());
         return isHovered(cursor, position, size);
     }
 
-    public static boolean isHovered(Vec2d cursor, Vec2d position, Vec2d size) {
+    public static boolean isHovered(Vec2i cursor, Vec2i position, Vec2i size) {
         return cursor.x() > position.x() && cursor.x() < position.x() + size.x() && cursor.y() > position.y() && cursor.y() < position.y() + size.y();
     }
 
-    public static boolean isHovered(Vec2d cursor, Vec2d position, Size size) {
+    public static boolean isHovered(Vec2i cursor, Vec2i position, Size size) {
         return cursor.x() > position.x() && cursor.x() < position.x() + size.x().pixel() && cursor.y() > position.y() && cursor.y() < position.y() + size.y().pixel();
     }
